+++ src/Generics/SOP/NP.hs
@@ -54,11 +54,20 @@ module Generics.SOP.NP
     -- * Collapsing
   , collapse_NP
   , collapse_POP
+    -- * Folding
+  , ctraverse__NP
+  , ctraverse__POP
+  , cfoldMap_NP
+  , cfoldMap_POP
     -- * Sequencing
   , sequence'_NP
   , sequence'_POP
   , sequence_NP
   , sequence_POP
+  , ctraverse'_NP
+  , ctraverse'_POP
+  , ctraverse_NP
+  , ctraverse_POP
     -- * Catamorphism and anamorphism
   , cata_NP
   , ccata_NP
@@ -77,6 +86,7 @@ module Generics.SOP.NP
 
 #if !(MIN_VERSION_base(4,8,0))
 import Control.Applicative
+import Data.Monoid (Monoid (..))
 #endif
 import Data.Coerce
 import Data.Proxy (Proxy(..))
@@ -501,21 +511,87 @@ type instance CollapseTo POP a = [[a]]
 instance HCollapse NP  where hcollapse = collapse_NP
 instance HCollapse POP where hcollapse = collapse_POP
 
+-- * Folding
+
+-- | Specialization of 'hctraverse_'.
+--
+-- @since 0.3.2.0
+--
+ctraverse__NP ::
+     forall c proxy xs f g. (All c xs, Applicative g)
+  => proxy c -> (forall a. c a => f a -> g ()) -> NP f xs -> g ()
+ctraverse__NP _ f = go
+  where
+    go :: All c ys => NP f ys -> g ()
+    go Nil       = pure ()
+    go (x :* xs) = f x *> go xs
+
+-- | Specialization of 'hcfoldMap'.
+--
+-- @since 0.3.2.0
+--
+ctraverse__POP ::
+     forall c proxy xs f g. (All2 c xs, Applicative g)
+  => proxy c -> (forall a. c a => f a -> g ()) -> POP f xs -> g ()
+ctraverse__POP p f = ctraverse__NP (allP p) (ctraverse__NP p f) . unPOP
+
+instance HTraverse_ NP  where hctraverse_ = ctraverse__NP
+instance HTraverse_ POP where hctraverse_ = ctraverse__POP
+
+-- | Specialization of 'hcfoldMap'.
+--
+-- @since 0.3.2.0
+--
+cfoldMap_NP :: (All c xs, Monoid m) => proxy c -> (forall a. c a => f a -> m) -> NP f xs -> m
+cfoldMap_NP  = hcfoldMap
+
+-- | Specialization of 'hcfoldMap'.
+--
+-- @since 0.3.2.0
+--
+cfoldMap_POP :: (All2 c xs, Monoid m) => proxy c -> (forall a. c a => f a -> m) -> POP f xs -> m
+cfoldMap_POP = hcfoldMap
+
 -- * Sequencing
 
 -- | Specialization of 'hsequence''.
-sequence'_NP  ::             Applicative f  => NP  (f :.: g) xs  -> f (NP  g xs)
+sequence'_NP  ::              Applicative f  => NP  (f :.: g) xs  -> f (NP  g xs)
 
 -- | Specialization of 'hsequence''.
 sequence'_POP :: (SListI xss, Applicative f) => POP (f :.: g) xss -> f (POP g xss)
+
+-- | Specialization of 'hctraverse''.
+--
+-- @since 0.3.2.0
+--
+ctraverse'_NP  ::
+     forall c proxy xs f f' g. (All c xs,  Applicative g)
+  => proxy c -> (forall a. c a => f a -> g (f' a)) -> NP f xs  -> g (NP f' xs)
+
+-- | Specialization of 'hctraverse''.
+--
+-- @since 0.3.2.0
+--
+ctraverse'_POP :: (All2 c xs, Applicative g) => proxy c -> (forall a. c a => f a -> g (f' a)) -> POP f xs -> g (POP f' xs)
 
 sequence'_NP Nil         = pure Nil
 sequence'_NP (mx :* mxs) = (:*) <$> unComp mx <*> sequence'_NP mxs
 
 sequence'_POP = fmap POP . sequence'_NP . hliftA (Comp . sequence'_NP) . unPOP
 
-instance HSequence NP  where hsequence' = sequence'_NP
-instance HSequence POP where hsequence' = sequence'_POP
+ctraverse'_NP _ f = go where
+  go :: All c ys => NP f ys -> g (NP f' ys)
+  go Nil       = pure Nil
+  go (x :* xs) = (:*) <$> f x <*> go xs
+
+ctraverse'_POP p f = fmap POP . ctraverse'_NP (allP p) (ctraverse'_NP p f) . unPOP
+
+instance HSequence NP  where
+  hsequence' = sequence'_NP
+  hctraverse' = ctraverse'_NP
+instance HSequence POP where
+  hsequence' = sequence'_POP
+  hctraverse' = ctraverse'_POP
 
 -- | Specialization of 'hsequence'.
 --
@@ -537,6 +613,21 @@ sequence_POP :: (All SListI xss, Applicative f) => POP f xss -> f (POP I xss)
 
 sequence_NP   = hsequence
 sequence_POP  = hsequence
+
+-- | Specialization of 'hctraverse'.
+--
+-- @since 0.3.2.0
+--
+ctraverse_NP  :: (All  c xs, Applicative g) => proxy c -> (forall a. c a => f a -> g a) -> NP  f xs -> g (NP  I xs)
+
+-- | Specialization of 'hctraverse'.
+--
+-- @since 0.3.2.0
+--
+ctraverse_POP :: (All2 c xs, Applicative g) => proxy c -> (forall a. c a => f a -> g a) -> POP f xs -> g (POP I xs)
+
+ctraverse_NP = hctraverse
+ctraverse_POP = hctraverse
 
 -- * Catamorphism and anamorphism